Create a date time utilities file which contains helper functions to handle date times.
Before this change we used methods provided by the `DateTimeProvider`
whenever we need to handle date times. This ensures that we always
consider the use timezone. In some cases we do not need to consider the
user timezone, e.g. when we want to

- transform a date time string to a date object, to substract e.g. a
  day.
- get the relative time for a date
- get the browser time

Using these helper functions instead of the context makes the code
simpler and easier to test.