Renaming `tsx` and `jsx` file extensions for files which do not implement JSX.
As described in the title, with this change we are renaming `tsx` and `jsx`
file extensions for files which do not implement JSX.
This not only helps to understand what a file is about when reading
its name. It also helps IDEs to process files better. E.g. when a file
and the related test file have different extensions, you can not switch
between them easily.