Improve handling of server errors when validating a query.
Currently we do not catch server errors when validating a search query,
which can affect the search bar validation and lead to unexpected behaviors.

We are now displaying a user notification and returning status: 'OK' in
this case. By returning this status we ensure we do not block the search
execution..