Try both session id and trusted header realms.
Prior to this change, only one of the authentication realms for session id
(including the session id extracted from a cookie) and trusted header
auth were executed exclusively for each request. This means that a
request that has a session cookie with a session id that does not match
the username set in the trusted header will lead to session termination,
but it cannot initiate a new session creation within the same request.

This change is now abstracting the authentication token to a shared
interface (`RemoteAddressAuthenticationToken`, as it was required to add
an additional shared field) which is implemented by both the session id
token as well as a dummy token used if none of the other authentication
mechanisms applies, effectively chaining the logic for session id and
trusted header when a session cookie is present. In this case, the logic
is:

  - extract session id from cookie
  - validate session id (is session present and authenticated?)
  - retrieve user for session id
  - if trusted header is present, do the usernames match?

If all of these hold true, authentication is completed for this session
id. If not, the trusted header authentication realm continue:

  - is trusted header auth enabled?
  - extract value of configured trusted header
  - is value present?
  - does the request come from a trusted subnet?

If these succeed, a new session will be created. When this is happening
through a session validation request, the resulting request will set a
new session cookie with this session id.