Adding date time utils.
This file contains helper functions to deal with dates in the
frontend. The goal is to have one place for our handling of dates, which
will make it easier to e.g. replace moment.

We will create a new context in the future which will use
these helper functions to provide dates based on the user time zone.