Creating date time utils. (#11986)
* Adding date time utils.

This file contains helper functions to deal with dates in the
frontend. The goal is to have one place for our handling of dates, which
will make it easier to e.g. replace moment.

We will create a new context in the future which will use
these helper functions to provide dates based on the user time zone.

* Avoid considering external time zome when using date time utils.

* Make it more clear which time zone will be used when using `DateTime` utils.

* Unify interface of date time utils.

* Remove not needed import.

* Updating comment.