Propagate output update with a new event type
Adds a new OutputChangeEvent which is propagated
to all cluster nodes via the cluster event bus.

This will lead to all cluster nodes restarting an
output after its configuration has changed.

Previously, the output was only restarted on the
node handling the UI request with the config update.
Other nodes in the cluster kept running the output
with a stale configuration.