Support credentials in URLs for MS Edge. (#12194)
Prior to this change, when generating URLs from the frontend that
directly link to the API (e.g. for downloading exports, reports, etc.),
it was checked if the browser supports embedding crendentials in the
URL. This check was written at a time when MS Edge was not based on
Chromium yet. Therefore it was assuming that it does not support it.

This has changed, therefore the check for Edge is removed. The whole
logic will vanish with the introduction of session cookies, therefore a
further revamp is not necessary.