Ensure that widget keeps id when moved to another dashboard page. (#12185) (#12214)
* Ensure that widget keeps id when moved to another dashboard page.

This change is making sure that when a widget is moved to a different dashboard page (without copying it), it will keep its id. This avoids any issues with referential integrity when the widget is referenced somewhere else by id (e.g. reporting).

Nothing changes when the widget is copied instead of being moved. The source widget will stay as is (keeping its id) and the new widget will get a new, random id. This way any references will stay intact, referencing the old widget.

* Replace `widget.id(uuid())` with `widget.newId()` which is doing the same.

Co-authored-by: Linus Pahl <linus@graylog.com>

Co-authored-by: Dennis Oelkers <dennis@graylog.com>