Implementing cookie authentication for API, using it in frontend. (#11788)
* Implementing cookie authentication for API, using it in frontend.

* Adding license headers.

* Also using cookie from request for proxied calls.

* Adding endpoint to terminate session without knowing session id.

* Cleaning up usages of session id.

* Adjusting test.

* Using UTF-8 for token.

* More cleanups.

* Removing `qualifyUrlWithSessionCredentials` as it is not needed anymore.

* Cleaning up tests.

* Removing `URLUtils.areCredentialsInURLSupported` and cleaning up.

* Cleaning up auth token creation, separating cookie auth token.

* Deprecating API method with session id path parameter.

* Using default version during cookie creation.

* Adding license headers.

* Adding integration test for authentication.

* Simplifying structure by reusing `SessionAuthenticator`.

* Including cookie when session is validated.

* Adding license header.

* Using username from response.

* Trigger validation of session after logging out.

* Centralizing creation of session for REST endpoints.

* Streamlining auth token creation.

* Consider `X-Graylog-Server-URL` as well when setting base URL of cookie.

* Deleting cookie when session is invalid or terminated.

* Setting session attribute for auth header.

* Try both session id and trusted header realms.

Prior to this change, only one of the authentication realms for session id
(including the session id extracted from a cookie) and trusted header
auth were executed exclusively for each request. This means that a
request that has a session cookie with a session id that does not match
the username set in the trusted header will lead to session termination,
but it cannot initiate a new session creation within the same request.

This change is now abstracting the authentication token to a shared
interface (`RemoteAddressAuthenticationToken`, as it was required to add
an additional shared field) which is implemented by both the session id
token as well as a dummy token used if none of the other authentication
mechanisms applies, effectively chaining the logic for session id and
trusted header when a session cookie is present. In this case, the logic
is:

  - extract session id from cookie
  - validate session id (is session present and authenticated?)
  - retrieve user for session id
  - if trusted header is present, do the usernames match?

If all of these hold true, authentication is completed for this session
id. If not, the trusted header authentication realm continue:

  - is trusted header auth enabled?
  - extract value of configured trusted header
  - is value present?
  - does the request come from a trusted subnet?

If these succeed, a new session will be created. When this is happening
through a session validation request, the resulting request will set a
new session cookie with this session id.

* Fixing `equals`/`hashCode` for `SessionIdToken`.

* Disabling linter hints.

* Explicitly omitting domain in cookie, leading to browser filling it in.

* Extracting `CookieFactory`.

* Cleaning up.

* Adding license header.

* Cleaning up unused session owner validation.

* Cleaning up.