Set lb status to "DEAD" earlier during shutdown
During shutdown, at first set the load balancer status to "DEAD" and then
pause for the configured load balancer recognition period. Only after that
trigger the lifecycle change to "HALTING".

Previously, we would start with triggering the lifecycle change to "HALTING"
prior to pausing for load balancer recognition. This would set the load
balancer status to "DEAD" but at the same time any service listening for the
lifecycle change to "HALTING" would already take action and potentially shut
down.

While in most cases this shouldn't be a problem, it is e.g. for the server-
side of the cloud forwarder which runs as a service and stops accepting
messages when server shutdown is detected. With this change, the load
balancer status is set now set to "DEAD" ahead of time.