Remove ancient artifact
It looks as this was required in an early version
when shutdown was triggered via a REST endpoint
and there was no lb grace period at all. Probably
the response to the initiating REST request was
in some cases not sent before the web server was
shut down.

I don't think we need this anymore