Fixing bug which resulted in multiple search creations and executions. (#12255)
* Fixing bug which resulted in multiple search creations and executions.

Before this change the deep equal check inside
`BindSearchParamsFromQuery` always returned false, because the filter of
the existing query was null and filter of the possible new query was
undefined. Of course, only when no query filter has been defined.

* Updating tests.

* Implementing previous solution to fix bug with search creation.

* Implementing better approach by adjusting `Query.equals` check.

This method will be called when checking the equality. Now it makes no
difference if a query `filter` is null or `undefined`.