Remove UI support for the GRAYLOG_API_URL environment variable
The GRAYLOG_API_URL environment variable was used for two development
features:

- Configuring the proxy backend for the development web server
  (devServer.js)
- Configuring the backend URL for the user interface (AppConfig.ts)

The latter will break cookie authentication when GRAYLOG_API_URL is set
because the URL might point to a different host than the web interface.
That prevents the cookie to be used for requests to the backend due to
the same-origin policy.

This commit removes the handling of the environment variable in the UI.

The development web server proxy backend can still be configured using
the variable.