Addresses an error when the enterprise plugin isn't installed. When a (#12475)
user navigates to the security page without the enterprise plugin
installed it will render correctly.
Before this when a user went to the security page it would error because
it was trying to use a link from the plugin routes. Now if the
enterprise plugin isn't installed it falls back to link to the
enterprise information page.