Fixing route prefixing after `urijs` update. (#12484)
* Fixing route prefixing after `urijs` update.

In #12468 `urijs` was updated. With this version, the parsing of URLs
starting with a double slash (e.g. `//search`) has changed, interpreting
the first path of the path as hostname (in this example `search`). This
led to application routes being messed up with when the application
prefix is `/`, swallowing the first part of the path, e.g. turning
`/gettingstarted` into `/`.

This change is now fixing the way application routes are being prefixed
and normalized. The `qualifyUrls` function in the `routing/Routes`
module is now returning proper results for all prefixes again, by making
use of the `URI#directory` function to set the directory path of a
prefixed URL, making sure that a double-slash is not parsed as part of a
scheme.

Although URL prefixing now works properly, a shortcut is added, which
just returns the original routes when the prefix is exactly `/`, saving
a few cycles.

Fixes #12476.