Ensure to not reset query string, time range and stream selection.
Before this change defining a value for a parameter resulted in a
reinitialization of the search bar form values. Because the parameter
bindings are being updated before we update the query, e.g. the previous query string
was being displayed during the update. Now we are memorizing the initial
values.