Always clear query validation explanations. (#12516)
* Always clear query validation explanations.

Before this change it was possible that old validation explanations
were not being removed. This happen because we used the error message as a key for the list items. Because of recent changes this attribute is no longer a uniq one.

* Defining better key for query explanation list items.

* Adding `id`s for validation explanations instead of relying on attributes for list item keys.

Before this change we weare using e.g. the error position attributes for
the list key. Since not all explanations have this information we need
to use an id instead.