Escape field value suggestions before adding them to query. (#12670)
Prior to this change, field value suggestions (existing field values
which are suggested in the query input when a user is wring a query)
were not escaped when the user selects one to be added to the query
string. This could lead to invalid queries being produced (e.g. when the
field contains a windows pathname).

This change is now adding escaping to those suggestions being produced
by the field value completion. A more general solution in the
`SearchBarAutoCompletions` class could be implemented at a later time,
but as this change aims towards being integrated into the upcoming
release, it tries to fix this issue in the least intrusive way.