Do not show error on messages from Forwarder in show single message page (#12485)
* Extract checking for local node from hook

* Refactor useMessage effect with async function

This will make the code a bit easier to read when changing logic to
check local node.

* Only load local inputs on show message page

On the show single message page, use `isLocalNode` function to decide
whether the message was received in a local node directly or through a
Forwarder. This way we can avoid failures when trying to fetch an input
from a local node, that actually belongs to a Forwarder.

Fixes https://github.com/Graylog2/graylog-plugin-enterprise/issues/3472