WIP concurrency limit for job scheduler
Implement a concurrency limit for each job definition.
The idea is to count the running job triggers for each job definition in
an aggregation, and exclude those from the nextRunnableTrigger() that
have reached their limit.

This needs to be done atomically!
The current solution uses transactions for this, which needs at least
Mongo 4.0 with a replica set.

Another possibility (instead of using transactions) might be to perform
the findOneAndUpdate() within the aggregation pipeline itself.
This needs mongo 4.2 at least:
 https://www.mongodb.com/docs/v4.2/tutorial/update-documents-with-aggregation-pipeline/
I haven't verified if this guarantees read consistency.

Changes:

 - Implement the nextRunnableTrigger() method without using mongojack.
   Mongojack does not yet support transactions.
   Without mongojack we can only use POJOs that can be converted to
   BSON. Thus we mainly use Strings and java.util.Date here.

- Support manual deserializing the JobTriggerDto with a jackson objectmapper.
  This would stumble over the "_id" field. Renaming it to "id" seems to
  be a solution.

- Change the MongoDBContainer to run with a replica set