Introduce JobResourceHandler / JobResourceHandlerService
Each JobResourceHandler implementation is responsible on checking the permissions of the UserContext
and perform the generic job actions accordingly.

All JobResourceHandlers will be called by the JobResourceHandlerService,
which can return either JobTriggerDTOs or for compat reasons with the
existing SystemJobs implementation: SystemJobSummary objects.

Duplicate Job type into the JobTriggerDTO.
This makes it easier to access that info when converting JobTriggers into
SystemJobSummary objects.

Add last_owner field to the JobTriggerDTO, so we can display which node
executed a job.