Detect fresh Graylog installations
This introduces a new named boolean that can be used to detect whether a fresh installation
of Graylog is happening.
It can be injected with: `@Named("isFreshInstallation")`

Knowing that a new installtion is happening allows us to introduce breaking changes.

Refactor MongoDBPreflightCheck so it can be used more than once.