Fix AddJobTypeToJobTrigger migration with Mongo 3.6
Mongo 3.6 does neither support $toString, nor $toObjectId
Thus we cannot use an aggregation to lookup the jobdefinition type.
We need to convert ObjectIds to Strings or vice versa to do the looukup.

Instead we are just fetching a map of all jobdefinition types.

Performance should still be okay, because we only expect lots of
JobTriggers, not JobDefinitions.