Introduce shared `Markdown` component.
This change is introducing a shared `Markdown` component, which solely
renders markdown passed in the `text` prop to HTML and returns it. The
reason for this is that we use this pattern repeatedly in an identical
way, using `dompurify` as an XSS filter, which can be easily forgotten.
Also, we have to include `marked` & `dompurify` in each plugin that
wants to do this.