cache assignments per sidecar id
using just the md5 of an assignment result might
leed to sidecars missing out on updates.
An assignment is always meant for a single sidecar, caching the entire
assignment over all sidecars is wrong.

A per sidecar cache entry allows more fine grained cache invalidations.

Bump the cacheMaxSize to 5000 entries, which is a more meaningful
number, considered that this might have to hold all sidecars.