Enforce usage of G1GC
There are some cases where Java falls back to using the
Serial GC instead of G1GC:
Less then 2GB of RAM or only one CPU.
This can be the case in some docker environments.
We always want to use G1GC, because it ensures that GC pauses are always short.

Also drop NewRatio=1 which was a tuning parameter used for the
previously used GC.
The default from Java is better here.
ResizeTLAB is also on by default now, so no need to specify that option.