Ensure Guice binder creation for the JobResourceHandler interface
We don't bind any JobResourceHandler classes in Graylog Open so we have
to make sure we create the binding.

This PR changes the RestApiBindings to extend PluginModule instead of
Graylog2Module to get access to the `jobResourceHandlerBinder()` method.

Fixes a Jersey injection error when navigating to "System/Overview" in a
pure-open Graylog instance.