Support more than 10 shutdown hooks in graceful shutdown service
The GracefulShutdownService used a SynchronousQueue with a fixed
capacity. If all threads were busy, the executor would reject
additional tasks.

We fix this by using an unbounded blocking queue for the executor.
We still limit the number of concurrently run hooks but allow
additional hooks to be queued for execution.