Add separate initialization step for outputs to handle license issues
Adds a separate initialization method to outputs which is called right
after constructing the instance. If an output requires a license or
has other requirements which are not met, it can throw an exception.
This prevents guice binding issues which might otherwise be logged if
the exception was thrown in the output's constructor.

Also improves logging for the special case of an output failing
because of an insufficient license.