Stop output only once when deleting globally
When an output was deleted, we would send an OutputDeletedEvent.
However, we would also remove the output from all streams, which
would emit a StreamsChangedEvent. Both events would be handled
by the OutputRegistry to stop the output. This lead to concurrent
stopping of the output which in case of journaled outputs could
cause issues when cleaning up the journal.

We can fix this by not emitting the OutputDeletedEvent because
the OutputRegistry can determine by the StreamsChangedEvent if
an output has to be deleted or not.