Simplify ULID creation
Use 32 bits instead of 16 bits for storing the sequence number.
This still leaves 48 bits of randomness in the ULID.
48 bits is still plenty of randomness to create unique ULIDs.
Considering that we only need them uniqe per timestamp and ES index.

With 32 bits we can store the seq as an unsigned int and don't have to worry
on how to shrink it down.
This simplifies the code a lot.