Only try to retain indices that are read only.
After the index rotation, we asynchronously start the
SetIndexReadOnlyJob, which will flush the index, set the index read only and
then set the closing Date.

This means that there is a time frame, where we can have rotated indices
that still haven't received their closing Date.

The retantion strategy would falsly assume that it is missing and
fallback to using the creation date.

Avoid this by fetching the IndicesBlockStatus and use it to filter
only indices that are read only ("index.blocks.write")

Add a test for the backwards compatble rention without a closing date.