Add aliases for IndexSetsDefaultConfiguration
The IndexSetConfig objects used a different naming scheme.

`rotation_strategy_config` is called `rotation_strategy`
and
`retention_strategy_config` is called `retention_strategy`

This is a problem for the UI code, which should be shared
on both pages.

Provide an alias for those values, otherwise we would need to write
a migration.