Add runtime failure tracking to polling inputs
Inputs that are polling their data, can experience intermittent failures.
So far, those were only logged as errors in the server.log.
There was no way to find these failures in the UI.

Introduce a new input IOState "Failing" which indicates those errors.

Add a new InputFailureRecorder to MessageInput.launch() which then can
be passed to the Transports of various Inputs, and replace
the LOG.errors certain calls.
The InputFailureRecorder can also be used to clear the failing input
state, once the error is resolved.

Inputs that want to make use of the new InputFailureRecorder
should inherit their Transports from ThrottleableTransport2,
which has the new launch/doLaunch signature.

In addition, the SystemNotification for failing Inputs could
not handle multiple Inputs at the time. To accomodate for this,
we introduce an optional "key" for SystemNotifications.
This allows us to differentiate between multiple Inputs.