Wrapping `latest` metric with filter clause for existence of field.
Prior to this change, the `latest(field)` metric did assume that `field`
is present in all documents of the result, otherwise it returns `N/A`.
This change is now adding a filter aggregation before the metric, so it
filters documents for the existence of the specified filter first and
returns the latest value found in the result set.

Fixes #13596.