Fix sun.reflect.Reflection.getCallerClass warning with Java >= 9
Enable the Multi-Release flag for the ManifestResourceTransformer in the
shade plugin. Without this, the JVM cannot take advantage of the class
files for newer JVM versions and hence logs the warning because slower
compatibility clases will be used at runtime.

Fixes the following warning on server startup:

    WARNING: sun.reflect.Reflection.getCallerClass is not supported.
    This will impact performance.

See: https://openjdk.org/jeps/238

Refs #7223
Refs #9802
Refs #11634