Fix OpenSearch permissions in data node tarball
We tried to use fileSet/fileMode to make the OpenSearch config
permissions work when running OpenSearch with a different user.

That didn't work because the fileMode changed all files and removed the
executable bit from all executable scripts.

Now we use the Maven antrun plugin to set the permissions on the
OpenSearch files right after extraction.