Fix global json-path library config modification in test class (#15225)
The "BooleanNumberConditionsVisitorTest" test was globally modifying the
configuration of the json-path library.

Since we are using two forked JVMs to run the tests, the changed
json-path configuration sometimes affected the "EncodingTest" test
class. That dependend on the distribution of the test files over the
forked JVMs and the order the tests are executed.

With surefire 2.22.1 the problem never showed up for some reason. With
surefire 3.0.0 the problem showed up more often that it did not. This
is probably due to the fork runner implementation changes in vesion 3.0.0.

This change also restores the trailing comma in the JSON example in
"EncodingTest" that got removed in the surefire update.