Add option to not replace "." with "_" in field names
Elasticsearch does not allow "." characters in field names since version 2.0.
Support has been restored since version 5.0.
For compatibility, graylog replaces "." with "_".
However, when Elasticsearch >= 5.0 is used, this is unnecessary.
For instance, Wazuh Indexer from the Wazuh project is forked from
Opensearch 1.3. The character replacement causes issues with Wazuh,
as the Dashboard expects dots as separator in the field name.

This adds the option `replace_dots_in_field_names` to revert this behavior and
allow the use of ".". The replacement is enabled by default for
compatibility with existing graylog configurations.

The extractor configuration in the web interface has been modified to show
a warning when the user inputs a "." in the "Key separator" field.

Closes: https://github.com/Graylog2/graylog2-server/issues/4583
Closes: https://github.com/Graylog2/graylog2-server/issues/6588
Closes: https://github.com/Graylog2/graylog2-server/issues/13043
Closes: https://github.com/Graylog2/graylog2-server/issues/14901
Bug: https://github.com/elastic/elasticsearch/issues/19443