Change log level of root logger to "info"
This is the de-facto standard, because during server startup
we are modifying the root logger level with "info" being the
default.

This happens in `CmdLineTool`. Before this change, commandsx
that don't inherit from `CmdLineTool` had to perform the same
kind of log level change. Now, they can just run with the
default "info" level which is what we probably want most of
the time.