Fix GELF output blocking and output processing race condition (#15501)
* Fix GELF output blocking and output processing race condition (#15385)

Switch from GelfTransport#send() to GelfTransport#trySend() to avoid
blocking output threads. We retry the send operation as long as the
output is running.
This fixes an issue where removing an output from a stream or deleting
it doesn't unblock processing. (with a slow output target)

A race condition in OutputRegistry was contributing to the problem.

When the OutputRegistry didn't have a cached output instance, we loaded
the output configuration from the database and started the output.
The problem is that each message holds a stream object that might be
outdated because the output could have been removed since adding the
stream object to the message object.

So instead of just starting an output, we need to check if the
output is still assigned to the stream.

Fixes #9917
Fixes Graylog2/graylog-plugin-enterprise#4623

(cherry picked from commit 30b9a912fde4f031ae04ad3a3d8983c9f7e616dc)

* Remove usage of rate limited logging