Fix join pipeline function parameter name
We cannot use `end` because it's a reserved keyword in the pipeline language and causes
a syntax error.

Instead of adding special handling for the `end` keyword to the pipeline rule language
we just rename the parameter to `endIndex`.

This shouldn't break existing rules, because they could've never been saved.

Fixes #11868

The substring function got already fixed by #15470 and should be
included in a backport.