Allow duplicating event definitions with a limit of 0
With #16035 we try to enforce a new default event limit,
without breaking backwards compatibility.

The validations are making an exemption if an existing
event definition is being edited which doesn't have an event limit yet
(limit of 0)

Duplicating event definitions was done in the frontend, and then
creating a new one.
This makes it impossible to differentiate between copy and create in the
backend.

 - Create a new `/events/definitions/{definitionId}/duplicate` endpoint.
   Which creates a copy without performing a validation.

 - Also correctly set the state for unscheduled event defintions.
   This got missed with #15558

Refs #16035