Introduce bootstrap configuration for plugin (#16205)
Bootstrap configuration can be part of a plugin
distribution but it will be loaded before the
actual plugin is instantiated. This allows to
e.g. specify a plugin-wide switch for a feature
which can then be evaluated early in the plugin
startup phase. If the feature is disabled, we
can skip loading configuration with potentially
required parameters and prevent validation
errors.

Co-authored-by: Laura Bergenthal-Grotlüschen <laura.bergenthalgrotlueschen@graylog.com>