Add execution duration to jobs overview
The duration info will be displayed as a tool-tip on the status badge.
Knowing how long it took a job (e.g. archive restore) to complete
can be helpful.

Furthermore, this info is persisted with the trigger in the DB, so we could also
build something like: show me the (current) slowest running event definition.