Move org.apache.directory.api:api-all into test scope
Adjust PipelineRuleParser and IndexFieldTypePollerPeriodical to use a
IdentityHashMap and ConcurrentHashMap instead of classes from mina-core.

Moving the api-all dependency to the test scope removes mina-core from
the compile scope. The IdentityHashSet and ConcurrentHashSet classes
from mina-core are basically small wrappers around IdentityHashMap and
ConcurrentHashMap. It's not worth including the mina-core dependency
just for those.