Record incoming stream metrics after processing
We are currently only recording the stream metrics in the
"Stream Rule Processor" stage.

Streams however, can be assigned from pipeline rules as well.

What makes that more tricky is the fact that they can also be
unassigned in pipeline rules.

The idea is to only record the metrics after all the processing
is done, and we know the final stream assignment.

Alternative fix for #17347

Fixes Graylog2/graylog-plugin-enterprise#6138