Reduce timerange for 'Show received messages' button
The default for a search when clicking on that button
was to search for all messages.
Executing this search can put a significant load on ES/OS and also
can lead to OOM situations.

Instead of searching all messages, we are searching only
messages in the past 5 minutes and the next future hour.

Fixes #5620
Fixes #4533
Fixes #10260