Add config to disable IPinfo database type check
The database type check is only a convenience feature to
give a meaningful error if the wrong type of mmdb is used.

In case the upstream metadata format changes again,
allow the check to be disabled temporarly.

The check can be disabled using:
`geo_ip_processor_disable_ipinfo_db_type_check=true`