Use gl2_second_sort_field alias for sorting
Indices created by Illuminate packages before 3.0
have created a field mapping for gl2_message_id of type text.

This broke our default sort order, which uses timestamp and
gl2_message_id as a tiebreaker.

To fix this, we do two things:

1. Patch the illuminate index templates
with an explicit keyword mapping for gl2_message_id.

2. Introduce a new alias `gl2_second_sort_field` which
points to `gl2_message_id` and will be used as a tiebreaker
from now on.

The fact that the new alias does not exist on older indices
allows the `unmapped` sort keyword to work again, and we can
keep using a default sort order with the tie breaker for
messages with identical timestamps.