Fix shutdown process of the job scheduler (#20141)
Let the JobSchedulerService register with the GracefulShutdownService to
ensure a correct shutdown order of the job scheduler components.

Previously, only the JobWorkerPool was registered in the
GracefulShutdownService. That caused the JobWorkerPool to be stopped
*before* the JobSchedulerService.

The JobExecutionContext#isShuttingDown method only returned true when
the JobSchedulerService was shut down. So jobs would only cancel their
execution when the scheduler service is in shutdown mode.

When the jobs are not cancelled, the JobWorkerPool shutdown didn't
finish and ran into a 60 second timeout. The JobSchedulerService was
only shutdown after the JobWorkerPool timeout.