Refactor InputRegistry to use a concurrent map for its state
The previous implementation using a concurrent set was prone to lose new
IOState objects because the MessageInput#equal and #hashCode where only
comparing the database ID. That lead to situations where a new IOState
object wasn't added to the set when an existing one was still present.

This change itself doesn't fix any issues, but it makes it more explicit
how we store the IOState in the map. (keyed by input ID) Using a map
also simplifies the access patterns to the state. (e.g., avoiding loops
to find an input by ID)