Remove async input starting from InputLauncher
We probably chose an async startup to decouple the startup from the
event bus thread and to start all inputs in parallel on server startup.

The first issue is solved since we now use the InputEventQueue which
already decouples the input startup from the event bus.

Removing complexity and cross-thread access to the input registry from
the input startup process is more important than having parallel input
starts.