Remove custom #hashCode and #equals methods from MessageInput
Both methods only took the database ID of the Input into acount. This
was probably implemented because the InputRegistry used to be a set.

The previous InputRegistry implementation cause much trouble and we
moved away from a set. So we can remove the custom methods here to avoid
problems in the future.