Fix display of encrypted configuration values for inputs (#21927)
Fix the display of encrypted input values that also have the IS_PASSWORD
attribute. We now first check if a field is an encrypted value before
checking the IS_PASSWORD status.

This fixes an issue where an empty encrypted input value was shown as
"****" instead of "<empty>" on the inputs page.

Also, add a space between the input label and the masked value for
password fields.