Don't go into journal recovery when shutting down (#22715)
* Don't go into journal recovery when shutting down

In 7a0b955f228e7b0f39ba2274e5c6af75532be691 we implemented a journal
recovery mechanism that tries to read from truncated journals. We
trigger the recovery when the #read() method returns an empty list.

The method can also return an empty list when the journal is in shutdown
mode. In that case we don't want to go into recovery mode.

* Add changelog

* Add a test to check the new behavior