Remove concurrency from dispatch-pr-build workflow
The PR dispatch workflow runs should not cancel each other!
When opening a PR with labels (e.g., dependabot), we receive one "opened"
and one or more "labeled" events. If we would cancel the workflow run
for "opened" when receiving a "labeled", the PR build in
graylog-project-internal would not be triggered.