Always store auth context in session
If a session auth context is provided during authentication, we will now always store it
to a session. If no session exists at this point, we will create a session because we
can treat the presence of a SessionAuthContext as a sufficient indicator that the
creation of a session is expected/required.

Previously, we were only storing the auth context in the session if the session existed.
Most likely because of a refactoring of the SessionCreator in
5cdac60877dc11fcdc81f4765cedbbc5b3f72971, the listener that persists the data is now
called before the session is created.